refactor: reorganize source modules and document internal APIs
Rename and reorganize the source tree:

- Move hashing, encryption, and key recovery from common.c into crypto.c.
- Move compression and decompression from extract.c into compression.c.
- Move generic codec-chain helpers from writer.c into compression.c.
- Keep archive lifecycle, sector buffering, and serialized I/O in writer.c.
- Update Automake sources and internal includes for the new module layout.

Add detailed function contracts for compression, crypto, Huffman, PKWARE,
reader, WAVE, and writer interfaces, including ownership, buffer handling,
return values, errors, and state requirements.

Document all regression-test helpers and test entry points, and verify the
updated source layout with formatting, compiler, test, and distribution checks.
