+++ src/Makefile.am
@@ -1,6 +1,6 @@
 # Build the shared library and keep private headers out of installations.
 lib_LTLIBRARIES			= libmpq.la
-noinst_HEADERS			= common.h crypt_buf.h endian.h extract.h huffman.h mpq-internal.h pack_begin.h pack_end.h pkware.h platform.h reader.h wave.h writer.h
+noinst_HEADERS			= compression.h crypt_buf.h crypto.h endian.h huffman.h mpq-internal.h pack_begin.h pack_end.h pkware.h platform.h reader.h wave.h writer.h
 
 AM_CPPFLAGS			= -I$(top_builddir) -I$(top_srcdir)/include
 
@@ -9,8 +9,8 @@ libmpq_la_LDFLAGS		= -version-info @LIBMPQ_ABI@
 
 # Library source files.
 GENERAL_SRCS = \
-	common.c \
-	extract.c \
+	compression.c \
+	crypto.c \
 	huffman.c \
 	mpq.c \
 	pkware.c \
