--- src/common.h
@@ -1,50 +1,0 @@
-/*
- *  common.h -- internal hash, crypt and decompression helpers.
- *
- *  Copyright (c) 2003-2026 Maik Broemme <mbroemme@libmpq.org>
- *
- *  This file is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- *  the Free Software Foundation; either version 2.1 of the License, or
- *  (at your option) any later version.
- *
- *  This file is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with this file; if not, see <https://www.gnu.org/licenses/>.
- */
-
-#ifndef LIBMPQ_COMMON_H
-#define LIBMPQ_COMMON_H
-
-#include <stdint.h>
-
-/* Hash an MPQ table name or file name with one of the Storm hash table offsets. */
-uint32_t libmpq__common_hash_string(const char *key, uint32_t offset);
-
-/* Encrypt a serialized little-endian block in place using the MPQ cipher. */
-int32_t libmpq__common_encrypt_block(uint8_t *in_buf, uint32_t in_size, uint32_t seed);
-
-/* Decrypt a serialized little-endian block in place using the MPQ cipher. */
-int32_t libmpq__common_decrypt_block(uint8_t *in_buf, uint32_t in_size, uint32_t seed);
-
-/* Recover a file seed from the first known words of an encrypted payload. */
-int32_t libmpq__common_detect_file_key(
-    const uint8_t *in_buf, uint32_t in_size, uint32_t file_size, uint32_t *key
-);
-
-/* Recover a file block-table decryption seed from encrypted block offsets. */
-int32_t libmpq__common_derive_block_table_seed(
-    uint8_t *in_buf, uint32_t in_size, uint32_t block_size, uint32_t *key
-);
-
-/* Decompress one archive block according to its MPQ compression flags. */
-int32_t libmpq__common_decompress_block(
-    uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size,
-    uint32_t compression_type
-);
-
-#endif /* LIBMPQ_COMMON_H */
