+++ src/compression.c
@@ -1,5 +1,5 @@
 /*
- *  extract.c -- decompression backends for MPQ block payloads.
+ *  compression.c -- MPQ compression and decompression orchestration.
  *
  *  Copyright (c) 2003-2026 Maik Broemme <mbroemme@libmpq.org>
  *
@@ -17,9 +17,10 @@
  *  along with this file; if not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "extract.h"
+#include "compression.h"
 #include "endian.h"
 #include "huffman.h"
+#include "mpq-internal.h"
 #include "pkware.h"
 #include "wave.h"
 #include <libmpq/mpq.h>
@@ -34,19 +35,19 @@
 static decompress_table_s dcmp_table[] = {
 
     /* Reverse of the canonical writer order. */
-    { LIBMPQ_COMPRESSION_BZIP2, libmpq__extract_decompress_bzip2 },
-    { LIBMPQ_COMPRESSION_PKZIP, libmpq__extract_decompress_pkzip },
-    { LIBMPQ_COMPRESSION_ZLIB, libmpq__extract_decompress_zlib },
-    { LIBMPQ_COMPRESSION_HUFFMAN, libmpq__extract_decompress_huffman },
-    { LIBMPQ_COMPRESSION_WAVE_STEREO, libmpq__extract_decompress_wave_stereo },
-    { LIBMPQ_COMPRESSION_WAVE_MONO, libmpq__extract_decompress_wave_mono }
+    { LIBMPQ_COMPRESSION_BZIP2, libmpq__compression_decompress_bzip2 },
+    { LIBMPQ_COMPRESSION_PKZIP, libmpq__compression_decompress_pkzip },
+    { LIBMPQ_COMPRESSION_ZLIB, libmpq__compression_decompress_zlib },
+    { LIBMPQ_COMPRESSION_HUFFMAN, libmpq__compression_decompress_huffman },
+    { LIBMPQ_COMPRESSION_WAVE_STEREO, libmpq__compression_decompress_wave_stereo },
+    { LIBMPQ_COMPRESSION_WAVE_MONO, libmpq__compression_decompress_wave_mono }
 };
 
 /* Decompress an MPQ Huffman-compressed stream into the caller-provided buffer.
  * The stream owns adaptive tree state, so the function initializes and frees
  * a separate tree and bit reader for each archive block. */
 int32_t
-libmpq__extract_decompress_huffman(
+libmpq__compression_decompress_huffman(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -93,7 +94,7 @@ libmpq__extract_decompress_huffman(
  * The output count returned by zlib is converted to the libmpq block API's
  * signed transfer convention, while zlib failures are propagated unchanged. */
 int32_t
-libmpq__extract_decompress_zlib(
+libmpq__compression_decompress_zlib(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -135,7 +136,7 @@ libmpq__extract_decompress_zlib(
  * A scratch codec object and callback state isolate the decoder from the
  * caller's buffers while preserving the exact number of produced bytes. */
 int32_t
-libmpq__extract_decompress_pkzip(
+libmpq__compression_decompress_pkzip(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -176,7 +177,7 @@ libmpq__extract_decompress_pkzip(
  * The bzip2 state consumes the compressed block and writes decoded bytes
  * directly to the destination supplied by the archive reader. */
 int32_t
-libmpq__extract_decompress_bzip2(
+libmpq__compression_decompress_bzip2(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -214,7 +215,7 @@ libmpq__extract_decompress_bzip2(
  * The channel count is fixed to one so the shared WAVE decoder can validate
  * the payload format and reconstruct the original PCM byte stream. */
 int32_t
-libmpq__extract_decompress_wave_mono(
+libmpq__compression_decompress_wave_mono(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -233,7 +234,7 @@ libmpq__extract_decompress_wave_mono(
  * The channel count is fixed to two, matching the stereo ADPCM framing and
  * predictor state expected by the shared WAVE decoder. */
 int32_t
-libmpq__extract_decompress_wave_stereo(
+libmpq__compression_decompress_wave_stereo(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -252,7 +253,7 @@ libmpq__extract_decompress_wave_stereo(
  * The leading mask selects supported codecs, and intermediate buffers preserve
  * each stage's output while the next stage consumes it. */
 int32_t
-libmpq__extract_decompress_multi(
+libmpq__compression_decompress_multi(
     uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size
 )
 {
@@ -331,5 +332,238 @@ libmpq__extract_decompress_multi(
 
     free(temp_buf);
 
+    return tb;
+}
+
+/* Return whether every requested compression bit has a local implementation.
+ * Unknown bits are rejected before any codec or archive state is modified. */
+int
+libmpq__compression_supported_mask(uint32_t mask)
+{
+    return (mask & ~(LIBMPQ_COMPRESSION_HUFFMAN | LIBMPQ_COMPRESSION_ZLIB |
+                     LIBMPQ_COMPRESSION_PKZIP | LIBMPQ_COMPRESSION_BZIP2 |
+                     LIBMPQ_COMPRESSION_WAVE_MONO | LIBMPQ_COMPRESSION_WAVE_STEREO)) == 0;
+}
+
+/* Apply one selected compression backend and replace the current buffer.
+ * Each backend receives the current stage output and returns a newly owned
+ * buffer, allowing the caller to retain the previous stage on fallback. */
+static int32_t
+compression_stage(uint8_t **data, size_t *size, uint32_t mask)
+{
+    uint8_t *out;
+    size_t out_size = *size + (*size / 100) + 1024;
+    z_stream z;
+    bz_stream b;
+    int result;
+
+    if (mask == LIBMPQ_COMPRESSION_HUFFMAN)
+        out_size = *size * 2 + 64;
+    out = malloc(out_size ? out_size : 1);
+    if (out == NULL)
+        return LIBMPQ_ERROR_MALLOC;
+    if (mask == LIBMPQ_COMPRESSION_PKZIP) {
+        uint8_t *candidate = NULL;
+        uint32_t candidate_size = 0;
+        int32_t status;
+        free(out);
+        status = libmpq__pkzip_compress(*data, (uint32_t)*size, &candidate, &candidate_size);
+        if (status < 0)
+            return status;
+        free(*data);
+        *data = candidate;
+        *size = candidate_size;
+        return LIBMPQ_SUCCESS;
+    } else if (mask == LIBMPQ_COMPRESSION_HUFFMAN) {
+        struct huffman_tree_s *tree = calloc(1, sizeof(*tree));
+        struct huffman_output_stream_s stream;
+        int32_t encoded;
+        if (tree == NULL) {
+            free(out);
+            return LIBMPQ_ERROR_MALLOC;
+        }
+        stream.out_buf = out;
+        stream.capacity = (uint32_t)out_size;
+        encoded = libmpq__huffman_encode(tree, &stream, *data, (uint32_t)*size);
+        free(tree);
+        if (encoded < 0) {
+            free(out);
+            return encoded;
+        }
+        free(*data);
+        *data = out;
+        *size = (size_t)encoded;
+        return LIBMPQ_SUCCESS;
+    } else if (mask == LIBMPQ_COMPRESSION_WAVE_MONO || mask == LIBMPQ_COMPRESSION_WAVE_STEREO) {
+        uint8_t *candidate = NULL;
+        uint32_t candidate_size = 0;
+        int32_t status = libmpq__wave_compress(
+            *data, (uint32_t)*size, &candidate, &candidate_size,
+            mask == LIBMPQ_COMPRESSION_WAVE_MONO ? 1 : 2
+        );
+        free(out);
+        if (status < 0)
+            return status;
+        free(*data);
+        *data = candidate;
+        *size = candidate_size;
+        return LIBMPQ_SUCCESS;
+    } else if (mask == LIBMPQ_COMPRESSION_ZLIB) {
+        memset(&z, 0, sizeof(z));
+        z.next_in = *data;
+        z.avail_in = (uInt)*size;
+        z.next_out = out;
+        z.avail_out = (uInt)out_size;
+        result = deflateInit(&z, Z_DEFAULT_COMPRESSION);
+        if (result == Z_OK)
+            result = deflate(&z, Z_FINISH);
+        if (result == Z_STREAM_END)
+            result = deflateEnd(&z);
+        else
+            deflateEnd(&z);
+        if (result != Z_OK) {
+            free(out);
+            return LIBMPQ_ERROR_UNPACK;
+        }
+        out_size = z.total_out;
+    } else {
+        memset(&b, 0, sizeof(b));
+        b.next_in = (char *)*data;
+        b.avail_in = (unsigned int)*size;
+        b.next_out = (char *)out;
+        b.avail_out = (unsigned int)out_size;
+        result = BZ2_bzCompressInit(&b, 9, 0, 30);
+        if (result == BZ_OK) {
+            do {
+                result = BZ2_bzCompress(&b, BZ_FINISH);
+            } while (result == BZ_FINISH_OK);
+        }
+        if (result == BZ_STREAM_END)
+            result = BZ2_bzCompressEnd(&b);
+        else
+            BZ2_bzCompressEnd(&b);
+        if (result != BZ_OK) {
+            free(out);
+            return LIBMPQ_ERROR_UNPACK;
+        }
+        out_size = b.total_out_lo32;
+    }
+    free(*data);
+    *data = out;
+    *size = out_size;
+    return LIBMPQ_SUCCESS;
+}
+
+/* Apply the selected compression chain and return its actual successful mask.
+ * Stages run in canonical Storm order, and a stage is kept only when it saves
+ * at least two bytes; the emitted mask therefore describes actual reductions. */
+int32_t
+libmpq__compression_encode_sector(
+    const uint8_t *input, size_t input_size, uint32_t requested, uint8_t **output,
+    size_t *output_size, uint8_t *emitted_mask
+)
+{
+    uint8_t *data;
+    size_t size;
+    uint32_t masks[] = { LIBMPQ_COMPRESSION_WAVE_MONO, LIBMPQ_COMPRESSION_WAVE_STEREO,
+                         LIBMPQ_COMPRESSION_HUFFMAN,   LIBMPQ_COMPRESSION_ZLIB,
+                         LIBMPQ_COMPRESSION_PKZIP,     LIBMPQ_COMPRESSION_BZIP2 };
+    size_t i;
+
+    if (!libmpq__compression_supported_mask(requested))
+        return LIBMPQ_ERROR_FORMAT;
+    data = malloc(input_size ? input_size : 1);
+    if (data == NULL)
+        return LIBMPQ_ERROR_MALLOC;
+    memcpy(data, input, input_size);
+    size = input_size;
+    *emitted_mask = 0;
+
+    /* Try each requested codec independently so expansion does not poison later stages. */
+    for (i = 0; i < sizeof(masks) / sizeof(masks[0]); i++) {
+        if (requested & masks[i]) {
+            size_t before = size;
+            uint8_t *saved = malloc(before ? before : 1);
+            int32_t result;
+            if (saved == NULL) {
+                free(data);
+                return LIBMPQ_ERROR_MALLOC;
+            }
+            memcpy(saved, data, before);
+            if (masks[i] == LIBMPQ_COMPRESSION_WAVE_MONO ||
+                masks[i] == LIBMPQ_COMPRESSION_WAVE_STEREO ||
+                masks[i] == LIBMPQ_COMPRESSION_HUFFMAN) {
+                free(saved);
+                continue;
+            }
+            result = compression_stage(&data, &size, masks[i]);
+            if (result < 0) {
+                free(saved);
+                continue;
+            }
+            if (size <= before - (before >= 2 ? 2 : before))
+                *emitted_mask |= (uint8_t)masks[i];
+            else {
+                free(data);
+                data = saved;
+                size = before;
+                saved = NULL;
+            }
+            free(saved);
+        }
+    }
+    if (*emitted_mask == 0 || size + 1 >= input_size) {
+        *emitted_mask = 0;
+        *output = data;
+        *output_size = size;
+    } else {
+        uint8_t *packed = realloc(data, size + 1);
+        if (packed == NULL) {
+            free(data);
+            return LIBMPQ_ERROR_MALLOC;
+        }
+        memmove(packed + 1, packed, size);
+        packed[0] = *emitted_mask;
+        *output = packed;
+        *output_size = size + 1;
+    }
+    return LIBMPQ_SUCCESS;
+}
+
+/* Decompress one archive block according to its MPQ compression flags.
+ * Raw data is copied directly, while PKWARE and multi-compression payloads
+ * are dispatched to the codec layer with MPQ-compatible expansion semantics. */
+int32_t
+libmpq__compression_decompress_block(
+    uint8_t *in_buf, uint32_t in_size, uint8_t *out_buf, uint32_t out_size,
+    uint32_t compression_type
+)
+{
+    int32_t tb = 0;
+
+    if (compression_type == LIBMPQ_FLAG_COMPRESS_NONE) {
+        if (in_size < out_size)
+            return LIBMPQ_ERROR_SIZE;
+        memcpy(out_buf, in_buf, out_size);
+        tb = out_size;
+    } else if (compression_type == LIBMPQ_FLAG_COMPRESS_PKZIP ||
+               compression_type == LIBMPQ_FLAG_COMPRESS_MULTI) {
+        if (compression_type == LIBMPQ_FLAG_COMPRESS_PKZIP) {
+            if (in_size >= out_size) {
+                memcpy(out_buf, in_buf, out_size);
+                tb = out_size;
+            } else if ((tb = libmpq__compression_decompress_pkzip(
+                            in_buf, in_size, out_buf, out_size
+                        )) < 0) {
+                return tb;
+            }
+        } else if (in_size < out_size) {
+            if ((tb = libmpq__compression_decompress_multi(in_buf, in_size, out_buf, out_size)) < 0)
+                return tb;
+        } else {
+            memcpy(out_buf, in_buf, out_size);
+            tb = out_size;
+        }
+    }
     return tb;
 }
