+++ src/writer.c
@@ -20,10 +20,9 @@
 #include "config.h"
 #endif
 
-#include "common.h"
+#include "compression.h"
+#include "crypto.h"
 #include "endian.h"
-#include "extract.h"
-#include "huffman.h"
 #include "mpq-internal.h"
 #include "pkware.h"
 #include "wave.h"
@@ -33,10 +32,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/stat.h>
-
-#include <bzlib.h>
-#include <zlib.h>
 
 /* Return the smallest supported power-of-two table capacity at or above value.
  * MPQ hash tables use power-of-two probing, so this helper provides the next
@@ -78,206 +73,9 @@ file_key(const char *name)
     for (i = 0; i < length; i++)
         normalized[i] = name[i] == '/' ? '\\' : name[i];
     normalized[length] = 0;
-    key = libmpq__common_hash_string(normalized, 0x300);
+    key = libmpq__crypto_hash_string(normalized, 0x300);
     free(normalized);
     return key;
-}
-
-/* Return whether every requested compression bit has a local implementation.
- * Unknown bits are rejected before any codec or archive state is modified. */
-static int
-supported_mask(uint32_t mask)
-{
-    return (mask & ~(LIBMPQ_COMPRESSION_HUFFMAN | LIBMPQ_COMPRESSION_ZLIB |
-                     LIBMPQ_COMPRESSION_PKZIP | LIBMPQ_COMPRESSION_BZIP2 |
-                     LIBMPQ_COMPRESSION_WAVE_MONO | LIBMPQ_COMPRESSION_WAVE_STEREO)) == 0;
-}
-
-/* Apply one selected compression backend and replace the current buffer.
- * Each backend receives the current stage output and returns a newly owned
- * buffer, allowing the caller to retain the previous stage on fallback. */
-static int32_t
-compress_stage(uint8_t **data, size_t *size, uint32_t mask)
-{
-    uint8_t *out;
-    size_t out_size = *size + (*size / 100) + 1024;
-    z_stream z;
-    bz_stream b;
-    int result;
-
-    if (mask == LIBMPQ_COMPRESSION_HUFFMAN)
-        out_size = *size * 2 + 64;
-    out = malloc(out_size ? out_size : 1);
-    if (out == NULL)
-        return LIBMPQ_ERROR_MALLOC;
-    if (mask == LIBMPQ_COMPRESSION_PKZIP) {
-        uint8_t *candidate = NULL;
-        uint32_t candidate_size = 0;
-        free(out);
-        int32_t status =
-            libmpq__pkzip_compress(*data, (uint32_t)*size, &candidate, &candidate_size);
-        if (status < 0)
-            return status;
-        free(*data);
-        *data = candidate;
-        *size = candidate_size;
-        return LIBMPQ_SUCCESS;
-    } else if (mask == LIBMPQ_COMPRESSION_HUFFMAN) {
-        struct huffman_tree_s *tree = calloc(1, sizeof(*tree));
-        struct huffman_output_stream_s stream;
-        int32_t encoded;
-        if (tree == NULL) {
-            free(out);
-            return LIBMPQ_ERROR_MALLOC;
-        }
-        stream.out_buf = out;
-        stream.capacity = (uint32_t)out_size;
-        encoded = libmpq__huffman_encode(tree, &stream, *data, (uint32_t)*size);
-        free(tree);
-        if (encoded < 0) {
-            free(out);
-            return encoded;
-        }
-        free(*data);
-        *data = out;
-        *size = (size_t)encoded;
-        return LIBMPQ_SUCCESS;
-    } else if (mask == LIBMPQ_COMPRESSION_WAVE_MONO || mask == LIBMPQ_COMPRESSION_WAVE_STEREO) {
-        uint8_t *candidate = NULL;
-        uint32_t candidate_size = 0;
-        int32_t status = libmpq__wave_compress(
-            *data, (uint32_t)*size, &candidate, &candidate_size,
-            mask == LIBMPQ_COMPRESSION_WAVE_MONO ? 1 : 2
-        );
-        free(out);
-        if (status < 0)
-            return status;
-        free(*data);
-        *data = candidate;
-        *size = candidate_size;
-        return LIBMPQ_SUCCESS;
-    } else if (mask == LIBMPQ_COMPRESSION_ZLIB) {
-        memset(&z, 0, sizeof(z));
-        z.next_in = *data;
-        z.avail_in = (uInt)*size;
-        z.next_out = out;
-        z.avail_out = (uInt)out_size;
-        result = deflateInit(&z, Z_DEFAULT_COMPRESSION);
-        if (result == Z_OK)
-            result = deflate(&z, Z_FINISH);
-        if (result == Z_STREAM_END)
-            result = deflateEnd(&z);
-        else
-            deflateEnd(&z);
-        if (result != Z_OK) {
-            free(out);
-            return LIBMPQ_ERROR_UNPACK;
-        }
-        out_size = z.total_out;
-    } else {
-        memset(&b, 0, sizeof(b));
-        b.next_in = (char *)*data;
-        b.avail_in = (unsigned int)*size;
-        b.next_out = (char *)out;
-        b.avail_out = (unsigned int)out_size;
-        result = BZ2_bzCompressInit(&b, 9, 0, 30);
-        if (result == BZ_OK) {
-            do {
-                result = BZ2_bzCompress(&b, BZ_FINISH);
-            } while (result == BZ_FINISH_OK);
-        }
-        if (result == BZ_STREAM_END)
-            result = BZ2_bzCompressEnd(&b);
-        else
-            BZ2_bzCompressEnd(&b);
-        if (result != BZ_OK) {
-            free(out);
-            return LIBMPQ_ERROR_UNPACK;
-        }
-        out_size = b.total_out_lo32;
-    }
-    free(*data);
-    *data = out;
-    *size = out_size;
-    return LIBMPQ_SUCCESS;
-}
-
-/* Apply the selected compression chain and return its actual successful mask.
- * Stages run in canonical Storm order, and a stage is kept only when it saves
- * at least two bytes; the emitted mask therefore describes actual reductions. */
-static int32_t
-encode_sector(
-    const uint8_t *input, size_t input_size, uint32_t requested, uint8_t **output,
-    size_t *output_size, uint8_t *emitted_mask
-)
-{
-    uint8_t *data;
-    size_t size;
-
-    uint32_t masks[] = { LIBMPQ_COMPRESSION_WAVE_MONO, LIBMPQ_COMPRESSION_WAVE_STEREO,
-                         LIBMPQ_COMPRESSION_HUFFMAN,   LIBMPQ_COMPRESSION_ZLIB,
-                         LIBMPQ_COMPRESSION_PKZIP,     LIBMPQ_COMPRESSION_BZIP2 };
-
-    size_t i;
-
-    if (!supported_mask(requested))
-        return LIBMPQ_ERROR_FORMAT;
-    data = malloc(input_size ? input_size : 1);
-    if (data == NULL)
-        return LIBMPQ_ERROR_MALLOC;
-    memcpy(data, input, input_size);
-    size = input_size;
-    *emitted_mask = 0;
-
-    /* Try each requested codec independently so expansion does not poison later stages. */
-    for (i = 0; i < sizeof(masks) / sizeof(masks[0]); i++) {
-        if (requested & masks[i]) {
-            size_t before = size;
-            uint8_t *saved = malloc(before ? before : 1);
-            int32_t result;
-            if (saved == NULL) {
-                free(data);
-                return LIBMPQ_ERROR_MALLOC;
-            }
-            memcpy(saved, data, before);
-            if (masks[i] == LIBMPQ_COMPRESSION_WAVE_MONO ||
-                masks[i] == LIBMPQ_COMPRESSION_WAVE_STEREO ||
-                masks[i] == LIBMPQ_COMPRESSION_HUFFMAN) {
-                free(saved);
-                continue;
-            }
-            result = compress_stage(&data, &size, masks[i]);
-            if (result < 0) {
-                free(saved);
-                continue;
-            }
-            if (size <= before - (before >= 2 ? 2 : before))
-                *emitted_mask |= (uint8_t)masks[i];
-            else {
-                free(data);
-                data = saved;
-                size = before;
-                saved = NULL;
-            }
-            free(saved);
-        }
-    }
-    if (*emitted_mask == 0 || size + 1 >= input_size) {
-        *emitted_mask = 0;
-        *output = data;
-        *output_size = size;
-    } else {
-        uint8_t *packed = realloc(data, size + 1);
-        if (packed == NULL) {
-            free(data);
-            return LIBMPQ_ERROR_MALLOC;
-        }
-        memmove(packed + 1, packed, size);
-        packed[0] = *emitted_mask;
-        *output = packed;
-        *output_size = size + 1;
-    }
-    return LIBMPQ_SUCCESS;
 }
 
 /* Compress, encrypt, and append the writer's currently buffered sector.
@@ -298,7 +96,7 @@ stream_flush_sector(mpq_writer_s *writer)
     if (writer->sector_index >= writer->block_count)
         return LIBMPQ_ERROR_SIZE;
     if ((writer->options.flags & LIBMPQ_FILE_FLAG_COMPRESS) != 0)
-        result = encode_sector(
+        result = libmpq__compression_encode_sector(
             writer->data, writer->data_size, requested, &packed, &packed_size, &emitted
         );
     else if ((writer->options.flags & LIBMPQ_FILE_FLAG_IMPLODE) != 0) {
@@ -318,7 +116,7 @@ stream_flush_sector(mpq_writer_s *writer)
 
     /* Payload encryption uses the file seed advanced by the sector number. */
     if (writer->options.flags & LIBMPQ_FILE_FLAG_ENCRYPTED)
-        libmpq__common_encrypt_block(
+        libmpq__crypto_encrypt_block(
             packed, (uint32_t)packed_size, file_key(writer->name) + writer->sector_index
         );
     if (writer->offsets)
@@ -376,7 +174,7 @@ stream_finish(mpq_writer_s *writer)
         for (i = 0; i <= writer->block_count; i++)
             libmpq__store_le32(table + i * 4, writer->offsets[i]);
         if (writer->options.flags & LIBMPQ_FILE_FLAG_ENCRYPTED)
-            libmpq__common_encrypt_block(table, (uint32_t)table_size, file_key(writer->name) - 1);
+            libmpq__crypto_encrypt_block(table, (uint32_t)table_size, file_key(writer->name) - 1);
         result = write_at(archive->fp, writer->payload_offset, table, table_size);
         free(table);
         if (result < 0)
@@ -476,8 +274,8 @@ finalize_archive(mpq_archive_s *a)
         libmpq__store_le16(raw + i * 16 + 10, a->mpq_hash[i].platform);
         libmpq__store_le32(raw + i * 16 + 12, a->mpq_hash[i].block_table_index);
     }
-    libmpq__common_encrypt_block(
-        raw, (uint32_t)bytes, libmpq__common_hash_string("(hash table)", 0x300)
+    libmpq__crypto_encrypt_block(
+        raw, (uint32_t)bytes, libmpq__crypto_hash_string("(hash table)", 0x300)
     );
     if (write_at(a->fp, a->mpq_header.hash_table_offset, raw, bytes) < 0) {
         free(raw);
@@ -496,8 +294,8 @@ finalize_archive(mpq_archive_s *a)
         libmpq__store_le32(raw + i * 16 + 8, a->mpq_block[i].unpacked_size);
         libmpq__store_le32(raw + i * 16 + 12, a->mpq_block[i].flags);
     }
-    libmpq__common_encrypt_block(
-        raw, (uint32_t)bytes, libmpq__common_hash_string("(block table)", 0x300)
+    libmpq__crypto_encrypt_block(
+        raw, (uint32_t)bytes, libmpq__crypto_hash_string("(block table)", 0x300)
     );
     if (write_at(a->fp, a->mpq_header.block_table_offset, raw, bytes) < 0) {
         free(raw);
@@ -698,8 +496,8 @@ libmpq__writer_file_begin(
         return LIBMPQ_ERROR_FORMAT;
     }
     if ((options->flags & LIBMPQ_FILE_FLAG_COMPRESS) &&
-        (!supported_mask(options->compression_first) ||
-         !supported_mask(options->compression_next))) {
+        (!libmpq__compression_supported_mask(options->compression_first) ||
+         !libmpq__compression_supported_mask(options->compression_next))) {
         free(w);
         return LIBMPQ_ERROR_FORMAT;
     }
