+++ src/writer.h
@@ -22,22 +22,35 @@
 
 #include <libmpq/mpq.h>
 
+/* Create a seekable archive and initialize its writer metadata from options. */
 int32_t libmpq__writer_archive_create(
     mpq_archive_s **out, const char *path, const mpq_archive_create_options_s *options
 );
+
+/* Begin one named file and return a stateful streaming writer for its payload. */
 int32_t libmpq__writer_file_begin(
     mpq_archive_s *archive, const char *name, libmpq__off_t size, const mpq_file_options_s *options,
     mpq_writer_s **out
 );
+
+/* Append bytes to the current file, buffering and flushing complete sectors. */
 int32_t libmpq__writer_file_write(mpq_writer_s *writer, const uint8_t *buffer, libmpq__off_t size);
+
+/* Finish the current file, write its sector offsets, and publish its block entry. */
 int32_t libmpq__writer_file_finish(mpq_writer_s *writer);
+
+/* Add a complete in-memory file using begin, write, and finish semantics. */
 int32_t libmpq__writer_file_add(
     mpq_archive_s *archive, const char *name, const uint8_t *data, libmpq__off_t size,
     const mpq_file_options_s *options
 );
+
+/* Read source from disk and add it as a named archive file. */
 int32_t libmpq__writer_file_add_path(
     mpq_archive_s *archive, const char *name, const char *source, const mpq_file_options_s *options
 );
+
+/* Write final tables, optional listfile, and the completed archive header. */
 int32_t libmpq__writer_finalize(mpq_archive_s *archive);
 
 #endif /* LIBMPQ_WRITER_H */
