+++ tests/test-archive.c
@@ -4,6 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+/* Exercise archive creation, metadata queries, opening, and closing. */
 int
 main(void)
 {
